Code quality cleanup: stdlib replacements, performance fixes, safety improvements
- Replace hand-rolled byte encoding with encoding/binary (db.go)
- Replace bubble sort with sort.Slice for RRF results (turbopuffer.go)
- Fix processFileMove to use embedding cache instead of re-embedding (handlers.go)
- Add http.MaxBytesReader 512MB limit to handleUpload (handlers.go)
- Batch SaveCachedEmbeddings into single multi-value INSERT (db.go)
- Use Python mimetypes.guess_type instead of hand-rolled dict (chunkers.py)
- Add case-insensitive extension handling to Go detectFileType (handlers.go)
- Remove dead init() function (turbopuffer.go)
- Add HTTP client timeouts: 300s CLI/Modal, 120s Turbopuffer
- Remove dead loop variable (handlers.go)
- Remove stub detectDirectoryMoves function (diff.go)
- Change Mtime from float64 to int64 for nanosecond precision (models.go)
- Add error logging to S3 Rename delete step (s3.go)

+++ pkg/models/models.go
@@ -28,7 +28,7 @@ type RootMetadata struct {
 type FileState struct {
 	Size        int64   `json:"size"`
 	ContentHash string  `json:"content_hash"`
-	Mtime       float64 `json:"mtime"`
+	Mtime       int64   `json:"mtime"`
 }
 
 // FileChangeStatus enumerates possible diff outcomes.
